--- app/js/controllers/MainController.js
@@ -1,9 +1,0 @@
-'use strict';
-
-import module from '../app';
-
-module.controller('MainController', ($scope) => {
-	$scope.name = 'Ira';
-});
-
-export default module;
