+++ app/js/index.js
@@ -1,13 +1,11 @@
 import 'bootstrap';
-import './controllers/MainController';
-import './controllers/BaseAdminController';
+import './controllers/admin/BaseAdminController';
+import './controllers/admin/LoginController';
 import './controllers/admin/AdminMainController';
 import './controllers/admin/AdminCarouselController';
 import './controllers/admin/AdminContactsController';
 import './controllers/admin/AdminGalleryController';
 import './controllers/admin/AdminUpcomingController';
-import './controllers/LoginController';
 
 import './factories/LoginFactory';
 import './factories/NotificationFactory';
-
